--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -8,6 +8,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Enhancements
 
+* Improve `pod install` performance for pods with exact file paths rather than glob patterns  
+  [Muhammed Yavuz Nuzumlalı](https://github.com/manuyavuz)
+  [#7473](https://github.com/CocoaPods/CocoaPods/pull/7473)
+
 * Add support for modular header search paths, include "legacy" support.  
   [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
   [#7412](https://github.com/CocoaPods/CocoaPods/pull/7412)