This PR dramatically improves `pod install` performance for pod specs using exact file path lists for file attributes rather than glob patterns.

## Changes

- `FileReferencesInstaller` was reading file system from scratch using `read_file_system` multiple times for a `PathList` object if multiple file accessors are using that same `PathList` object. This is a common case for development pods having several subspecs which are independently added into a `Podfile`.
Now, file system will be read only once for each `PathList` object in `refresh_file_accessors` method.

- When `PathList.relative_glob` method runs, all string objects in `patterns` argument are regarded as glob patterns and `File.fnmatch` runs for each `(path_list_file, pattern)` pairs, which has complexity O(m*n). 
Now, we calculate the set intersection of list of all files exist under `PathList` and list of patterns, to find exact file path matches much faster. Then, we run `File.fnmatch` for remaining glob patterns.

Here is `time` output of my `pod install` run before and after changes:

*Before :* `bundle exec pod install  51.63s user 4.46s system 99% cpu 56.524 total`
*After :* `bundle exec pod install  7.08s user 1.97s system 98% cpu 9.176 total`