@segiddins 2018-03-05
This doesn't seem to help our `pod install` times at all :(