@manuyavuz 2018-03-05
Do you use array of exact file paths for attributes such as `source_files` or `public_header_files`?

In our project, we were having a PathList object for our development pod, and it had `40005` as number of files in it. And number of patterns given to `relative_glob` method was `446`, all of which were exact file paths. So we get rid of 40005x446 `File.fnmatch` calls.

If PathList root does not contain too many files or `podspec` file does not have many exact file paths, this change won't provide any improvements.

Maybe our case is a bit extreme, but changes does not have any measurable impact on projects using glob patterns, so it's good to have it :)

Merging.