@dnkoutso 2018-03-05
We mostly use a single glob as we've standardized our file structure. This is why we do not see as much gains in performance from this. We did once try on a pod to specify exactly the files and it made it super-slow. Your PR fixes this which is great.